+++ truck-geometry/src/decorators/rbf_surface/mod.rs
@@ -61,10 +61,8 @@ pub trait RadiusFunction: Clone {
     fn der2(&self, t: f64) -> f64 { self.der_n(2, t) }
     /// Substitutes the higher-order derivations to `out`.
     #[inline]
-    fn ders(&self, t: f64, out: &mut [f64]) {
-        out.iter_mut()
-            .enumerate()
-            .for_each(|(i, o)| *o = self.der_n(i, t))
+    fn ders(&self, max_order: usize, t: f64) -> CurveDers<f64> {
+        (0..=max_order).map(|n| self.der_n(n, t)).collect()
     }
 }
 
@@ -149,13 +147,15 @@ where
 {
     type Point = Point3;
     type Vector = Vector3;
-    fn ders<A: AsMut<[Vector3]>>(&self, u: f64, v: f64, out: &mut [A]) {
+    fn ders(&self, max_order: usize, u: f64, v: f64) -> SurfaceDers<Vector3> {
         let cc = self.contact_circle(v).unwrap();
-        out.iter_mut().enumerate().for_each(|(i, out)| {
-            out.as_mut().iter_mut().enumerate().for_each(|(j, o)| {
-                *o = self.sub_der_mn(i, j, u, cc);
+        let mut out = SurfaceDers::new(max_order);
+        (0..=max_order).for_each(|i| {
+            (0..=max_order - i).for_each(|j| {
+                out[i][j] = self.sub_der_mn(i, j, u, cc);
             });
         });
+        out
     }
     fn der_mn(&self, m: usize, n: usize, u: f64, v: f64) -> Vector3 {
         self.sub_der_mn(m, n, u, self.contact_circle(v).unwrap())
@@ -276,8 +276,7 @@ where
             return self.subs(t).to_vec();
         }
         let cc = self.surface.contact_circle(t).unwrap();
-        let mut rders = [0.0; 32];
-        self.surface.radius.ders(t, &mut rders[0..=n]);
+        let rders = self.surface.radius.ders(n, t);
         let cc_ders = self.surface.center_contacts_ders(cc, &rders, n);
         match self.index {
             0 => cc_ders.contact0_ders[n],